+++ infinite-rs/src/lib.rs
@@ -31,7 +31,7 @@ fn load_modules() -> Result<()> {
 ## Loading a tag file
 After we have loaded a module file, we can now use the [`read_tag`](`ModuleFile::read_tag`) function to load a specific tag by index from the module file. This populates the [`data_stream`](`crate::module::file::ModuleFileEntry::data_stream`) and [`tag_info`](`crate::module::file::ModuleFileEntry::tag_info`) properties in a module entry that we can use later.
 
-The [`read_tag_from_id`](`ModuleFile::read_tag_from_id`) function is also available to load a tag by its global ID, returning the index in which it was found in the module file.
+The [`read_tag_from_id`](`ModuleFile::read_tag_from_id`) function is also available to load a tag by its global ID.
 
 ```rust
 use infinite_rs::{ModuleFile, Result};
@@ -41,11 +41,11 @@ fn load_tags() -> Result<()> {
 
     // Load a specific tag from the module file.
     let tag_index = 0;
-    let id = module.read_tag(tag_index)?;
-    if id.is_some() {
+    let tag = module.read_tag(tag_index)?;
+    if let Some(tag) = tag {
         // We can now access the data stream and tag info.
-        let tag_data = module.files[tag_index as usize].data_stream.as_ref().unwrap();
-        let tag_info = module.files[tag_index as usize].tag_info.as_ref().unwrap();
+        let tag_data = tag.data_stream.as_ref().unwrap();
+        let tag_info = tag.tag_info.as_ref().unwrap();
     }
     Ok(())
 }
@@ -106,14 +106,14 @@ fn load_tags() -> Result<()> {
     // And for each material tag, we want to read the metadata associated.
     for index in material_indices {
         // We first have to populate data_stream and tag_info.
-        let id = module.read_tag(index as u32)?;
-        if id.is_some() {
+        let tag = module.read_tag(index as u32)?;
+        if let Some(tag) = tag {
             let mut mat = MaterialTag::default();
             // We pass in our structure as a generic parameter.
-            module.files[index].read_metadata(&mut mat)?;
+            tag.read_metadata(&mut mat)?;
             // We can now access the fields in our structure.
             // For instance, `any_tag.internal_struct.tag_id` is always equal to the tag id of our file.
-            assert_eq!(module.files[index].tag_id, mat.any_tag.internal_struct.tag_id);
+            assert_eq!(tag.tag_id, mat.any_tag.internal_struct.tag_id);
         }
     }
     Ok(())